NOJIRA  Prevent updates to existing nodes from overwriting protected properties (such as IDs and paths)
This issue came up when the LiteJsonImporter was processing JSON such as:

  {
      "parent": {
          "child": {
              "_id": "somerubbish",
              "foo": "bar"
          },
          "somekey": "value",
          "_id": "morerubbish"
      }
  }

The LiteJsonImporter walks this tree depth first and creates the child nodes before the parent nodes, but when child nodes are created, stub nodes get created for their parents.  When processing returns to the top-level and goes to set the properties for the "parent" node, the node is already there and we're dealing with an "update" rather than an "insert".

ContentManagerImpl would stop newly-created content nodes from providing their own values for protected fields, but it didn't do the same check for updates, and since LiteJsonImporter's created nodes are always updated, this meant that nodes would take on whatever garbage was in the source JSON data.